+++ CotEditor/Sources/CETextView.h
@@ -41,7 +41,6 @@ extern NSString *_Nonnull const CETextViewDidBecomeFirstResponderNotification;
 @property (nonatomic) BOOL needsRecompletion;  // 再度入力補完をするか
 @property (nonatomic) BOOL usesAntialias;
 @property (nonatomic) BOOL showsInvisibles;
-@property (nonatomic) CGFloat lineSpacing;
 @property (nonatomic) NSUInteger tabWidth;  // タブ幅
 @property (nonatomic) NSRect highlightLineRect;  // ハイライト行の矩形
 @property (nonatomic, getter=isAutoTabExpandEnabled) BOOL autoTabExpandEnabled;  // タブを自動的にスペースに展開するか