+++ CotEditor/Sources/CETextViewProtocol.h
@@ -9,7 +9,7 @@
 
  ------------------------------------------------------------------------------
  
- © 2014-2015 1024jp
+ © 2014-2016 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -32,7 +32,6 @@
 @protocol CETextViewProtocol
 
 @required
-- (CGFloat)lineSpacing;
 - (nullable CETheme *)theme;
 
 @end